Fix issues after checking
